Consolidate duplicated test seed helpers into conftests
The nauro-core operations-kernel tests each carried a near-identical
_seed_decision / _store_with pair, and the search and embeddings tests
duplicated a _make_decision builder byte for byte. The cross-surface
parity tests likewise repeated _seed_filesystem_store / _seed_cloud_store
and a _dump shim across modules. Lifting these into conftests removes the
drift risk without changing any test's observable behavior: the shared
helpers are supersets whose defaults preserve every call site's produced
store shape and seeded content.

nauro-core: add tests/conftest.py hosting union _seed_decision, _store_with,
and make_decision; migrate the five operations tests plus test_search and
test_embeddings_union to import them.

nauro: add tests/cross_surface/conftest.py with parameterized
seed_filesystem_store / seed_cloud_store and the shared _dump; migrate the
six seeded modules and the three _dump users. CloudStore stays a lazy import
so the conftest loads even when the private mcp_server package is absent, and
each module keeps its own importorskip gate.

The deliberately divergent seeders (the get_context and decision_lookup
full-store fixtures, and the test_context / test_graph decision builders) are
left in place.
