+++ packages/nauro-core/tests/operations/test_check_decision.py
@@ -10,6 +10,7 @@ from __future__ import annotations
 from datetime import date
 
 import pytest
+from conftest import _seed_decision, _store_with
 
 from nauro_core.constants import (
     LEXICAL_RANK_CAVEAT,
@@ -18,12 +19,7 @@ from nauro_core.constants import (
     NO_DECISIONS_TO_CHECK,
     NO_RELATED_DECISIONS,
 )
-from nauro_core.decision_model import (
-    Decision,
-    DecisionConfidence,
-    DecisionStatus,
-    format_decision,
-)
+from nauro_core.decision_model import DecisionStatus
 from nauro_core.operations import (
     CheckDecisionResult,
     InMemoryStore,
@@ -31,34 +27,6 @@ from nauro_core.operations import (
 )
 from nauro_core.operations.check_decision import _assessment
 from nauro_core.operations.results import RelatedDecision
-
-
-def _seed_decision(
-    num: int,
-    title: str,
-    rationale: str,
-    *,
-    status: DecisionStatus = DecisionStatus.active,
-    decision_date: date | None = None,
-) -> tuple[str, str]:
-    """Return (file_stem, formatted_markdown) for a minimal v2 decision."""
-    superseded_by = "999" if status is DecisionStatus.superseded else None
-    decision = Decision(
-        date=decision_date or date(2026, 1, 1),
-        confidence=DecisionConfidence.medium,
-        status=status,
-        superseded_by=superseded_by,
-        num=num,
-        title=title,
-        rationale=rationale,
-    )
-    slug = title.lower().replace(" ", "-")
-    stem = f"{num:03d}-{slug}"
-    return stem, format_decision(decision)
-
-
-def _store_with(*decisions: tuple[str, str]) -> InMemoryStore:
-    return InMemoryStore(decisions=dict(decisions))
 
 
 def test_returns_result_type() -> None:
