+++ packages/nauro-core/tests/operations/test_get_decision.py
@@ -9,6 +9,8 @@ from __future__ import annotations
 
 from datetime import date
 
+from conftest import _seed_decision, _store_with
+
 from nauro_core.decision_model import (
     Decision,
     DecisionConfidence,
@@ -23,35 +25,6 @@ from nauro_core.operations import (
     get_decision,
 )
 from nauro_core.operations.get_decision import _LEDE_MAX_CHARS
-
-
-def _seed_decision(
-    num: int,
-    title: str,
-    rationale: str,
-    *,
-    status: DecisionStatus = DecisionStatus.active,
-    stem: str | None = None,
-) -> tuple[str, str]:
-    """Return (file_stem, formatted_markdown) for a minimal v2 decision."""
-    superseded_by = "999" if status is DecisionStatus.superseded else None
-    decision = Decision(
-        date=date(2026, 1, 1),
-        confidence=DecisionConfidence.medium,
-        status=status,
-        superseded_by=superseded_by,
-        num=num,
-        title=title,
-        rationale=rationale,
-    )
-    if stem is None:
-        slug = title.lower().replace(" ", "-")
-        stem = f"{num:03d}-{slug}"
-    return stem, format_decision(decision)
-
-
-def _store_with(*decisions: tuple[str, str]) -> InMemoryStore:
-    return InMemoryStore(decisions=dict(decisions))
 
 
 def test_returns_result_type() -> None:
